More modifications, scrap some deprecated XML
