--- src/CSF.Core/Utils/Extensions/AsyncExtensions.cs
@@ -1,22 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace CSF
-{
-    internal static class AsyncExtensions
-    {
-        public static ValueTask<T> AsValueTask<T>(this T value)
-        {
-            return new ValueTask<T>(value);
-        }
-
-        public static ValueTask<T> AsValueTask<T>(this Task<T> task)
-        {
-            return new ValueTask<T>(task);
-        }
-
-        public static ValueTask AsValueTask(this Task task)
-        {
-            return new ValueTask(task);
-        }
-    }
-}
